Fixing a couple of failed tests on php 7.4 snapshot
Have seen these failed tests for a while, decided to look into it.

Apparently in php 7.4 it's no longer possible to use the `current` php function on an object, and expect it to use the Traversable methods. Instead it traverses the public properties of the object (which is expected behavior in most cases).

More here: https://github.com/php/php-src/blob/07df6594b5c71b176b376aba8007bc33616bbfb0/UPGRADING#L75-L86
And more discussion in a bug report here: https://bugs.php.net/bug.php?id=77864

I decided to leave the code alone and just fix the tests in this case, as any code changes to resolve this are a BC break, and the code itself doesn't cause issues in PHP 7.4, but the way the tests were interfacing with the code was the issue. Technically a `getIterator` method should probably return an iterator, not an `ArrayObject` that has to then have `getIterator` called on it, but this change shouldn't break any implementations using this code (that code may break on php 7.4 though if using `current`, `next`, etc... on the return of `getIterator` directly).

Also fixing a couple of minor docblock issues that came up from testing a code analysis tool.